% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_to_draw.R
\name{obs_to_draw}
\alias{obs_to_draw}
\title{percentile}
\usage{
obs_to_draw(obs, draws)
}
\arguments{
\item{obs}{obs}

\item{draws}{comparisons}
}
\value{
percentile (percent of draws <= obs)
}
\description{
percentile
}
