% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_to_forest.R
\name{obs_to_forest_z}
\alias{obs_to_forest_z}
\title{Compared observed to random forest}
\usage{
obs_to_forest_z(hill1, s, n)
}
\arguments{
\item{hill1}{obs}

\item{s}{s}

\item{n}{n}
}
\value{
z score of obs | feasible set as predicted by rf
}
\description{
Compared observed to random forest
}
