% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_to_norm.R
\name{obs_to_norm_p}
\alias{obs_to_norm_p}
\title{P value}
\usage{
obs_to_norm_p(obs, mean, sd)
}
\arguments{
\item{obs}{obs}

\item{mean}{mean}

\item{sd}{sd}
}
\value{
p value of obs | N(mean, sd)
}
\description{
P value
}
