% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_to_norm.R
\name{obs_to_norm_z}
\alias{obs_to_norm_z}
\title{Z score}
\usage{
obs_to_norm_z(obs, mean, sd)
}
\arguments{
\item{obs}{obs}

\item{mean}{mean}

\item{sd}{sd}
}
\value{
z score of obs compared to N(mean, sd)
}
\description{
Z score
}
