% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_forest.R
\name{predict_forest}
\alias{predict_forest}
\title{Predict parameters from fitted random forest}
\usage{
predict_forest(s, n)
}
\arguments{
\item{s}{s}

\item{n}{n}
}
\value{
df
}
\description{
Predict parameters from fitted random forest
}
